restore config to run GoogleTests, upgrade some tests, ignore obsolete

+++ MBSolver/Tests/distancetest.cpp
@@ -8,7 +8,8 @@ TEST(distance,basic) {
 	int64_t biggest = std::numeric_limits<int64_t>::max( );
 	int64_t okay = DefaultDistancePolicy<int64_t>::convert(biggest);
 	ASSERT_TRUE(okay == biggest);
-	double tooBig = static_cast<double>(biggest) + 1;
+    // jcs: switched test to long double, a double cannot hold int64_t max+1 without underflow, using long double
+	long double tooBig = static_cast<long double>(biggest) + 1;
 	ASSERT_THROW(DefaultDistancePolicy<int64_t>::convert(tooBig),std::domain_error);
 	cout << okay << endl;
 }
